added operational automcompletions from r session - lots of cleanup needed
