+++ renderer/src/editor/editor.ts
@@ -4,11 +4,9 @@ import { initDoc } from "../session/sessionApi"
 import {EditorView} from "@codemirror/view"
 
 import {setup} from "./setup"
-import {rlang} from "../../argonaut-lezer-r/src"
+import {rlangsupport} from "../../argonaut-lezer-r/src"
 
 import { repdoc } from "../repdoc/repdoc"
-import { acTest1 } from "./autocomplete/acTest1"
-import { acTest2 } from "./autocomplete/acTest2"
 
 import { Editor } from "../appTypes"
 
@@ -30,7 +28,7 @@ export function getEditor(tabState: TabState, data: string, element: HTMLDivElem
             setup,
             //images(), //ignore the name - this does a console log printout of the parse tree
             repdoc({docSessionId: tabState.id}),
-            rlang({autocomplete: [acTest1,acTest2]})
+            rlangsupport()
         ],
         parent: element
     })
